requirements.txt
fastapi
uvicorn
pycryptodome
qrcode[pil]
requests
ascon
qiskit
qiskit-aer
sympy